tests/wpt/css-tests/css-shapes-1_dev/html/shape-outside-invalid-ellipse-006.htm

<!DOCTYPE html>
<html><head>
	<meta charset="UTF-8">
	<meta content="dom" name="flags">
    <meta content="shape-outside declaration is invalid if cx in ellipse function is invalid" name="assert">
	<title>CSS Shapes Test - Ellipse - invalid cx value</title>
	<link href="mailto:noonnightstorm@gmail.com" rel="author" title="noonnightstorm">
	<link href="mailto:stearns@adobe.com" rel="reviewer" title="Alan Stearns">
	<link href="http://www.w3.org/TR/css-shapes-1/#supported-basic-shapes" rel="help">
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<link href="/resources/testharness.css" rel="stylesheet">

	<style>
#shape{
	width: 300px;
	height: 300px;
	shape-outside: ellipse(50, 50%, 50%, 50%);
}
	</style>
</head>
<body>

	<div id="log"></div>

	<div id="shape"></div>

	<script>
	var obj = document.getElementById('shape');
	var ellipse = getComputedStyle(obj).shapeOutside;
	test(function() {assert_equals(ellipse, "none", "declaration should be Invalid")});
	</script>


</body></html>
